pyproject.toml
# Modern Python project configuration
# https://packaging.python.org/en/latest/specifications/pyproject-toml/

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "dataconverter"
version = "2.0.0"
description = "Professional CLI tool for converting data between JSON, YAML, TOML, and XML formats"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["data", "converter", "json", "yaml", "toml", "xml", "cli"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Utilities",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

# Runtime dependencies
dependencies = [
    "lark>=1.1.9",           # Parser generator for query language
    "jsonpath-ng>=1.6.1",    # JSONPath expressions
    "pyyaml>=6.0.1",         # YAML support
    "xmltodict>=0.13.0",     # XML to dict conversion
    "typer>=0.12.0",         # CLI framework
    "rich>=13.7.0",          # Beautiful terminal output
]

# Optional development dependencies
[project.optional-dependencies]
dev = [
    "black>=24.0.0",         # Code formatter
    "ruff>=0.6.0",           # Fast Python linter
    "mypy>=1.11.0",          # Type checker
]

# CLI entry point
[project.scripts]
dataconv = "main:main"

# Tool configurations
[tool.black]
line-length = 88
target-version = ["py310", "py311", "py312", "py313"]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by black)
]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
check_untyped_defs = true
strict_equality = true

# Ignore missing imports for third-party libraries
[[tool.mypy.overrides]]
module = ["jsonpath_ng.*", "xmltodict.*"]
ignore_missing_imports = true
